Always use LEFT JOIN when the referred-to table comes from a LEFT JOIN relation
In "a LEFT JOIN b INNER JOIN c", the last jointure can reject results that
were previously accepted by the LEFT JOIN. This is the case if a is not null,
b is null and c is matched against b. The correct result is to have b and c
null, but keeping the inner join will find no c matching the null b, and thus
will remove the entire row.